create top level dirs relayer, core, reportingplugin
