+++ pkg/loop/ccip_execution.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/smartcontractkit/chainlink-common/pkg/logger"
 	"github.com/smartcontractkit/chainlink-common/pkg/loop/internal/goplugin"
-	"github.com/smartcontractkit/chainlink-common/pkg/loop/internal/relayer/pluginprovider/ext/ccip"
+	"github.com/smartcontractkit/chainlink-common/pkg/loop/internal/reportingplugin/ccip"
 	"github.com/smartcontractkit/chainlink-common/pkg/types"
 )
