+++ pkg/loop/internal/core/services/reportingplugin/ocr2/reporting.go
