+++ pkg/loop/internal/core/services/reportingplugin/ocr3/reporting.go
