+++ pkg/loop/internal/core/services/reportingplugin/ocr3/reporting_plugin_service.go
