+++ pkg/loop/internal/core/services/reportingplugin/ocr3/test/factory.go
