+++ pkg/loop/internal/core/services/reportingplugin/ocr3/test/factory_generators.go
