+++ pkg/loop/reportingplugins/ocr3/grpc_test.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/smartcontractkit/chainlink-common/pkg/logger"
 	"github.com/smartcontractkit/chainlink-common/pkg/loop"
-	ocr3_test "github.com/smartcontractkit/chainlink-common/pkg/loop/internal/ocr3/test"
+	ocr3test "github.com/smartcontractkit/chainlink-common/pkg/loop/internal/core/services/reportingplugin/ocr3/test"
 	"github.com/smartcontractkit/chainlink-common/pkg/loop/internal/test"
 	testcore "github.com/smartcontractkit/chainlink-common/pkg/loop/internal/test/core"
 	"github.com/smartcontractkit/chainlink-common/pkg/types"
@@ -35,7 +35,7 @@ func PluginGenericTest(t *testing.T, p types.OCR3ReportingPluginClient) {
 			types.CapabilitiesRegistry(nil))
 		require.NoError(t, err)
 
-		ocr3_test.OCR3ReportingPluginFactory(t, factory)
+		ocr3test.OCR3ReportingPluginFactory(t, factory)
 	})
 }
 
@@ -45,9 +45,9 @@ func TestGRPCService_MedianProvider(t *testing.T) {
 	stopCh := newStopCh(t)
 	test.PluginTest(
 		t,
-		ocr3_test.OCR3ReportingPluginWithMedianProviderName,
+		ocr3test.OCR3ReportingPluginWithMedianProviderName,
 		&GRPCService[types.MedianProvider]{
-			PluginServer: ocr3_test.MedianServer,
+			PluginServer: ocr3test.MedianServer,
 			BrokerConfig: loop.BrokerConfig{
 				Logger: logger.Test(t),
 				StopCh: stopCh,
@@ -65,7 +65,7 @@ func TestGRPCService_PluginProvider(t *testing.T) {
 		t,
 		PluginServiceName,
 		&GRPCService[types.PluginProvider]{
-			PluginServer: ocr3_test.AgnosticPluginServer,
+			PluginServer: ocr3test.AgnosticPluginServer,
 			BrokerConfig: loop.BrokerConfig{
 				Logger: logger.Test(t),
 				StopCh: stopCh,
